+++ packages/ui/src/components/Pagination/__stories__/index.stories.tsx
@@ -17,5 +17,7 @@ export default {
 export { Playground } from './Playground.stories'
 export { Basic } from './Basic.stories'
 export { Disabled } from './Disabled.stories'
+export { Size } from './Size.stories'
 export { Controlled } from './Controlled.stories'
+export { HideButtons } from './HideButtons.stories'
 export { PerPage } from './PerPage.stories'