+++ packages/ui/src/compositions/index.ts
@@ -13,3 +13,18 @@ export { RichTextInput } from './RichTextInput'
 export { OrderSummary } from './OrderSummary'
 export { Plans } from './Plans'
 export { SteppedListCard } from './SteppedListCard'
+export { Filters, useFiltersContext, FiltersProvider, FiltersMainRow, FiltersDrawer } from './Filters'
+export type {
+  FiltersProps,
+  FilterConfig,
+  FilterConfigGroup,
+  FilterConfigItemBase,
+  FilterConfigItem,
+  FilterConfigItemDatetimeRange,
+  FilterConfigItemMultiSelect,
+  FilterConfigItemNumber,
+  FilterConfigItemSelect,
+  FilterConfigItemSlider,
+  FilterConfigItemText,
+  FilterComponentProps,
+} from './Filters'